+++ src/test/java/jezzsantos/automate/plugin/infrastructure/services/cli/AutomateCliServiceTests.java
@@ -69,16 +69,19 @@ public class AutomateCliServiceTests {
         var filename = createTemporaryFile(executableName);
         Mockito.when(this.configuration.getExecutablePath())
           .thenReturn(StringWithDefault.fromValue(filename));
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("anerror", ""));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult<>(new StructuredError("anerror"), null));
         Mockito.when(this.upgrader.upgrade(anyString(), any(), anyString(), any(), any()))
           .thenReturn(new CliExecutableStatus(executableName, "100.0.0"));
 
         new AutomateCliService(this.configuration, this.cache, this.platform, this.cliRunner, this.upgrader);
 
         Mockito.verify(this.configuration).getExecutablePath();
-        Mockito.verify(this.cliRunner).execute(argThat(x -> x.equals("aninstallationdirectory")), argThat(x -> x.equals(StringWithDefault.fromValue(filename))),
-                                               argThat(x -> x.size() == 1 && x.get(0).equals("--version")));
+        Mockito.verify(this.cliRunner)
+          .executeStructured(any(), argThat(x -> x.equals("aninstallationdirectory")), argThat(x -> x.equals(StringWithDefault.fromValue(filename))), anyBoolean(),
+                             argThat(x -> x.size() == 1
+                               && x.get(0).equals("info")
+                             ));
         Mockito.verify(this.cache).setIsCliInstalled(true);
         Mockito.verify(this.cliRunner).log(argThat(x -> x.Type == CliLogEntryType.NORMAL));
         Mockito.verify(this.upgrader)
@@ -94,16 +97,19 @@ public class AutomateCliServiceTests {
         var filename = createTemporaryFile(executableName);
         Mockito.when(this.configuration.getExecutablePath())
           .thenReturn(StringWithDefault.fromValue(filename));
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("", "0.0.0"));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult<>(null, new GetInfoStructuredOutput("0.0.0")));
         Mockito.when(this.upgrader.upgrade(anyString(), any(), anyString(), any(), any()))
           .thenReturn(new CliExecutableStatus(executableName, "100.0.0"));
 
         new AutomateCliService(this.configuration, this.cache, this.platform, this.cliRunner, this.upgrader);
 
         Mockito.verify(this.configuration).getExecutablePath();
-        Mockito.verify(this.cliRunner).execute(argThat(x -> x.equals("aninstallationdirectory")), argThat(x -> x.equals(StringWithDefault.fromValue(filename))),
-                                               argThat(x -> x.size() == 1 && x.get(0).equals("--version")));
+        Mockito.verify(this.cliRunner)
+          .executeStructured(any(), argThat(x -> x.equals("aninstallationdirectory")), argThat(x -> x.equals(StringWithDefault.fromValue(filename))), anyBoolean(),
+                             argThat(x -> x.size() == 1
+                               && x.get(0).equals("info")
+                             ));
         Mockito.verify(this.cache).setIsCliInstalled(true);
         Mockito.verify(this.cliRunner).log(argThat(x -> x.Type == CliLogEntryType.NORMAL));
         Mockito.verify(this.upgrader)
@@ -118,16 +124,19 @@ public class AutomateCliServiceTests {
         var filename = createTemporaryFile(this.service.getExecutableName());
         Mockito.when(this.configuration.getExecutablePath())
           .thenReturn(StringWithDefault.fromValue(filename));
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("", "100.0.0"));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult<>(null, new GetInfoStructuredOutput("100.0.0")));
         Mockito.when(this.upgrader.upgrade(anyString(), any(), anyString(), any(), any()))
           .thenAnswer(x -> x.getArgument(3));
 
         new AutomateCliService(this.configuration, this.cache, this.platform, this.cliRunner, this.upgrader);
 
         Mockito.verify(this.configuration).getExecutablePath();
-        Mockito.verify(this.cliRunner).execute(argThat(x -> x.equals("aninstallationdirectory")), argThat(x -> x.equals(StringWithDefault.fromValue(filename))),
-                                               argThat(x -> x.size() == 1 && x.get(0).equals("--version")));
+        Mockito.verify(this.cliRunner)
+          .executeStructured(any(), argThat(x -> x.equals("aninstallationdirectory")), argThat(x -> x.equals(StringWithDefault.fromValue(filename))), anyBoolean(),
+                             argThat(x -> x.size() == 1
+                               && x.get(0).equals("info")
+                             ));
         Mockito.verify(this.cache).setIsCliInstalled(true);
         Mockito.verify(this.cliRunner).log(argThat(x -> x.Type == CliLogEntryType.NORMAL));
         Mockito.verify(this.upgrader, never()).upgrade(anyString(), any(), anyString(), any(), any());
@@ -172,8 +181,8 @@ public class AutomateCliServiceTests {
     public void whenTryGetExecutableStatusAndInvalidExecutable_ThenReturnsStatus() {
 
         var filename = createTemporaryFile(this.service.getExecutableName());
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("anerror", ""));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult<>(new StructuredError("anerror"), null));
 
         var result = this.service.tryGetExecutableStatus("acurrentdirectory", StringWithDefault.fromValue(filename));
 
@@ -184,8 +193,8 @@ public class AutomateCliServiceTests {
     public void whenTryGetExecutableStatusAndUnsupportedVersion_ThenReturnsStatus() {
 
         var filename = createTemporaryFile(this.service.getExecutableName());
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("", "0.0.1"));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult<>(null, new GetInfoStructuredOutput("0.0.1")));
 
         var result = this.service.tryGetExecutableStatus("acurrentdirectory", StringWithDefault.fromValue(filename));
 
@@ -215,8 +224,8 @@ public class AutomateCliServiceTests {
         var filename = createTemporaryFile(this.service.getExecutableName());
         Mockito.when(this.configuration.getExecutablePath())
           .thenReturn(StringWithDefault.fromValue(filename));
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("", "0.0.1"));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult(null, new GetInfoStructuredOutput("0.0.1")));
         Mockito.when(this.cache.isCliInstalled(any()))
           .thenAnswer((Answer) invocation -> ((Supplier<Boolean>) invocation.getArguments()[0]).get());
 
@@ -232,8 +241,8 @@ public class AutomateCliServiceTests {
         var filename = createTemporaryFile(this.service.getExecutableName());
         Mockito.when(this.configuration.getExecutablePath())
           .thenReturn(StringWithDefault.fromValue(filename));
-        Mockito.when(this.cliRunner.execute(anyString(), any(), anyList()))
-          .thenReturn(new CliTextResult("", "100.0.0"));
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
+          .thenReturn(new CliStructuredResult(null, new GetInfoStructuredOutput("100.0.0")));
         Mockito.when(this.cache.isCliInstalled(any()))
           .thenAnswer((Answer) invocation -> ((Supplier<Boolean>) invocation.getArguments()[0]).get());
 
@@ -246,7 +255,7 @@ public class AutomateCliServiceTests {
     @Test
     public void whenListAllAutomationAndNotCachedAndFails_ThenReturnsEmptyLists() {
 
-        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyList()))
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
           .thenReturn(new CliStructuredResult<>(new StructuredError(), null));
         Mockito.when(this.cache.listAll(any(), anyBoolean()))
           .thenAnswer((Answer) invocation -> ((Supplier<AllStateLite>) invocation.getArguments()[0]).get());
@@ -258,7 +267,7 @@ public class AutomateCliServiceTests {
         assertTrue(result.getDrafts().isEmpty());
         Mockito.verify(this.cliRunner).executeStructured(
           any(), argThat(x -> x.equals("acurrentdirectory")),
-          argThat(x -> x.equals(StringWithDefault.fromValue("anexecutablepath"))),
+          argThat(x -> x.equals(StringWithDefault.fromValue("anexecutablepath"))), anyBoolean(),
           argThat(list -> list.equals(Arrays.asList("list", "all"))));
     }
 
@@ -266,7 +275,7 @@ public class AutomateCliServiceTests {
     @Test
     public void whenListAllAutomationAndNotCached_ThenReturnsCliResult() {
 
-        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyList()))
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
           .thenReturn(new CliStructuredResult(null, new ListAllDefinitionsStructuredOutput()));
         Mockito.when(this.cache.listAll(any(), anyBoolean()))
           .thenAnswer((Answer) invocation -> ((Supplier<AllStateLite>) invocation.getArguments()[0]).get());
@@ -278,14 +287,14 @@ public class AutomateCliServiceTests {
         assertTrue(result.getDrafts().isEmpty());
         Mockito.verify(this.cliRunner).executeStructured(
           any(), argThat(x -> x.equals("acurrentdirectory")),
-          argThat(x -> x.equals(StringWithDefault.fromValue("anexecutablepath"))),
+          argThat(x -> x.equals(StringWithDefault.fromValue("anexecutablepath"))), anyBoolean(),
           argThat(list -> list.equals(Arrays.asList("list", "all"))));
     }
 
     @Test
     public void whenListAllAutomationAndCached_ThenReturnsCached() {
 
-        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyList()))
+        Mockito.when(this.cliRunner.executeStructured(any(), anyString(), any(), anyBoolean(), anyList()))
           .thenReturn(new CliStructuredResult<>(new StructuredError(), null));
         Mockito.when(this.cache.listAll(any(), anyBoolean()))
           .thenReturn(new AllStateLite());
@@ -295,7 +304,7 @@ public class AutomateCliServiceTests {
         assertTrue(result.getPatterns().isEmpty());
         assertTrue(result.getToolkits().isEmpty());
         assertTrue(result.getDrafts().isEmpty());
-        Mockito.verify(this.cliRunner, never()).executeStructured(any(), anyString(), any(), anyList());
+        Mockito.verify(this.cliRunner, never()).executeStructured(any(), anyString(), any(), anyBoolean(), anyList());
     }
 
     @SuppressWarnings({"rawtypes", "unchecked"})